+++ src/gana/operators/composition.py
