+++ tests/sum_test.py
@@ -1,6 +1,6 @@
 import pytest
 from src.gana.block.program import Prg
-from src.gana.operations.operators import sigma
+from src.gana.operators.sigma import sigma
 from src.gana.sets.index import I
 from src.gana.sets.variable import V
 
@@ -125,7 +125,6 @@ def test_sigma(p):
         [p.v[105], p.v[106], p.v[107], p.v[108], p.v[109]],
         [p.v[110], p.v[111], p.v[112], p.v[113], p.v[114]],
     ]
-    assert [[j[n] for j in p.f5.variables] for n in range(len(p.f5))] == [[p.w[0], p.w[1], p.w[2], p.w[3], p.w[4]]]
-
-
-
+    assert [[j[n] for j in p.f5.variables] for n in range(len(p.f5))] == [
+        [p.w[0], p.w[1], p.w[2], p.w[3], p.w[4]]
+    ]
